Changes to Linux routing setup
This commit removes an existing IPv6 route if IPv6 is not enabled. This
ensures that traffic does not bypass the proxy via IPv6.

Further, this commit does not create the 0.0.0.0/1 and 128.0.0.0/1
routes if there is no default route. In this case, a default route is
created. This way, it does not interfere with OpenVPN in a new network
namespace without routes.